Introduce new API: completion-deferred-completions
Introduce deferred hilit functions and use them.

* lisp/minibuffer.el (completion--nth-completion): Add DEFER argument.
(completion--deferred-hilit): New function.
(completion-basic-all-completions): Use it.
(completion-pcm-all-completions): Use it.
(completion--pcm-deferred-hilit): New function.
(completion-flex-all-completions): Use it.
(completion-substring-all-completions): Use it.
(completion-pcm-all-completions): Use it.

[[ALTERNATIVE: Add DEFER argument to completion-all-completions]]

Remove the function `completion-deferred-completions`. The downside of
this approach is that we will carry the additional DEFER argument
around forever, such that we opt-in to the deferred highlighting and
the new return value format.  A completely new API
`completion-deferred-completions` can return an extensible result as
an alist/plist, which would guarantee that we don't have to introduce
another API again, due to a change of the return value.  We may still
add further optional arguments, but this is a backward compatible API
extension.

Also defer 22 and 21